@@ -1,4 +1,4 @@
-package cachekv
+package internal
 
 import (
 	"bytes"
@@ -18,17 +18,20 @@ type cacheMergeIterator struct {
 	parent    types.Iterator
 	cache     types.Iterator
 	ascending bool
+
+	valid bool
 }
 
 var _ types.Iterator = (*cacheMergeIterator)(nil)
 
-func newCacheMergeIterator(parent, cache types.Iterator, ascending bool) *cacheMergeIterator {
+func NewCacheMergeIterator(parent, cache types.Iterator, ascending bool) *cacheMergeIterator {
 	iter := &cacheMergeIterator{
 		parent:    parent,
 		cache:     cache,
 		ascending: ascending,
 	}
 
+	iter.valid = iter.skipUntilExistsOrInvalid()
 	return iter
 }
 
@@ -56,42 +59,38 @@ func (iter *cacheMergeIterator) Domain() (start, end []byte) {
 
 // Valid implements Iterator.
 func (iter *cacheMergeIterator) Valid() bool {
-	return iter.skipUntilExistsOrInvalid()
+	return iter.valid
 }
 
 // Next implements Iterator
 func (iter *cacheMergeIterator) Next() {
-	iter.skipUntilExistsOrInvalid()
 	iter.assertValid()
 
-	// If parent is invalid, get the next cache item.
-	if !iter.parent.Valid() {
+	switch {
+	case !iter.parent.Valid():
+		// If parent is invalid, get the next cache item.
 		iter.cache.Next()
-		return
-	}
-
-	// If cache is invalid, get the next parent item.
-	if !iter.cache.Valid() {
+	case !iter.cache.Valid():
+		// If cache is invalid, get the next parent item.
 		iter.parent.Next()
-		return
+	default:
+		// Both are valid.  Compare keys.
+		keyP, keyC := iter.parent.Key(), iter.cache.Key()
+		switch iter.compare(keyP, keyC) {
+		case -1: // parent < cache
+			iter.parent.Next()
+		case 0: // parent == cache
+			iter.parent.Next()
+			iter.cache.Next()
+		case 1: // parent > cache
+			iter.cache.Next()
+		}
 	}
-
-	// Both are valid.  Compare keys.
-	keyP, keyC := iter.parent.Key(), iter.cache.Key()
-	switch iter.compare(keyP, keyC) {
-	case -1: // parent < cache
-		iter.parent.Next()
-	case 0: // parent == cache
-		iter.parent.Next()
-		iter.cache.Next()
-	case 1: // parent > cache
-		iter.cache.Next()
-	}
+	iter.valid = iter.skipUntilExistsOrInvalid()
 }
 
 // Key implements Iterator
 func (iter *cacheMergeIterator) Key() []byte {
-	iter.skipUntilExistsOrInvalid()
 	iter.assertValid()
 
 	// If parent is invalid, get the cache key.
@@ -122,7 +121,6 @@ func (iter *cacheMergeIterator) Key() []byte {
 
 // Value implements Iterator
 func (iter *cacheMergeIterator) Value() []byte {
-	iter.skipUntilExistsOrInvalid()
 	iter.assertValid()
 
 	// If parent is invalid, get the cache value.
@@ -153,11 +151,12 @@ func (iter *cacheMergeIterator) Value() []byte {
 
 // Close implements Iterator
 func (iter *cacheMergeIterator) Close() error {
+	err1 := iter.cache.Close()
 	if err := iter.parent.Close(); err != nil {
 		return err
 	}
 
-	return iter.cache.Close()
+	return err1
 }
 
 // Error returns an error if the cacheMergeIterator is invalid defined by the
