@@ -1,20 +1,32 @@
-package streaming
+package streaming_test
 
 import (
 	"testing"
 
+	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/codec"
 	codecTypes "github.com/cosmos/cosmos-sdk/codec/types"
+	serverTypes "github.com/cosmos/cosmos-sdk/server/types"
+	"github.com/cosmos/cosmos-sdk/simapp"
+	"github.com/cosmos/cosmos-sdk/store/streaming"
 	"github.com/cosmos/cosmos-sdk/store/streaming/file"
 	"github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/tendermint/tendermint/libs/log"
+	dbm "github.com/tendermint/tm-db"
 
 	"github.com/stretchr/testify/require"
 )
 
 type fakeOptions struct{}
 
-func (f *fakeOptions) Get(string) interface{} { return nil }
+func (f *fakeOptions) Get(key string) interface{} {
+	if key == "streamers.file.write_dir" {
+		return "data/file_streamer"
+
+	}
+	return nil
+}
 
 var (
 	mockOptions       = new(fakeOptions)
@@ -24,12 +36,12 @@ var (
 )
 
 func TestStreamingServiceConstructor(t *testing.T) {
-	_, err := NewServiceConstructor("unexpectedName")
+	_, err := streaming.NewServiceConstructor("unexpectedName")
 	require.NotNil(t, err)
 
-	constructor, err := NewServiceConstructor("file")
+	constructor, err := streaming.NewServiceConstructor("file")
 	require.Nil(t, err)
-	var expectedType ServiceConstructor
+	var expectedType streaming.ServiceConstructor
 	require.IsType(t, expectedType, constructor)
 
 	serv, err := constructor(mockOptions, mockKeys, testMarshaller)
@@ -39,5 +51,57 @@ func TestStreamingServiceConstructor(t *testing.T) {
 	for _, key := range mockKeys {
 		_, ok := listeners[key]
 		require.True(t, ok)
+	}
+}
+
+func TestLoadStreamingServices(t *testing.T) {
+	db := dbm.NewMemDB()
+	encCdc := simapp.MakeTestEncodingConfig()
+	keys := sdk.NewKVStoreKeys("mockKey1", "mockKey2")
+	bApp := baseapp.NewBaseApp("appName", log.NewNopLogger(), db, nil)
+
+	testCases := map[string]struct {
+		appOpts            serverTypes.AppOptions
+		activeStreamersLen int
+	}{
+		"empty app options": {
+			appOpts: simapp.EmptyAppOptions{},
+		},
+		"all StoreKeys exposed": {
+			appOpts:            streamingAppOptions{keys: []string{"*"}},
+			activeStreamersLen: 1,
+		},
+		"some StoreKey exposed": {
+			appOpts:            streamingAppOptions{keys: []string{"mockKey1"}},
+			activeStreamersLen: 1,
+		},
+		"not exposing anything": {
+			appOpts: streamingAppOptions{keys: []string{"mockKey3"}},
+		},
+	}
+
+	for name, tc := range testCases {
+		t.Run(name, func(t *testing.T) {
+			activeStreamers, _, err := streaming.LoadStreamingServices(bApp, tc.appOpts, encCdc.Marshaler, keys)
+			require.NoError(t, err)
+			require.Equal(t, tc.activeStreamersLen, len(activeStreamers))
+		})
+	}
+}
+
+type streamingAppOptions struct {
+	keys []string
+}
+
+func (ao streamingAppOptions) Get(o string) interface{} {
+	switch o {
+	case "store.streamers":
+		return []string{"file"}
+	case "streamers.file.keys":
+		return ao.keys
+	case "streamers.file.write_dir":
+		return "data/file_streamer"
+	default:
+		return nil
 	}
 }
